fix: ensure update-manpages service respects proxy
The `update-manpages` systemd unit runs a
collection of scripts, some of which invoke tools
such as `curl` which interact with the network.

While `curl` will respect proxy config vars such
as `HTTP_PROXY` by default, it won't natively
understand `JUJU_CHARM_HTTP_PROXY` or similar.

This patch ensures that if those variables are
set when the charm executes, the respective
configuration is added to the systemd unit.